pyproject.toml
[tool.poetry]
name = "warehouse-ia"
version = "0.1.0"
description = ""
authors = ["Igor Martins <igormartins0301@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "3.5.2"
requests = "^2.32.3"
boto3 = "^1.35.20"
python-dotenv = "^1.0.1"
delta-spark = "3.2.0"
taskipy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
pre_test = 'task lint'
post_test = 'coverage html'